Classes.EvalSteps

Description

Documentation

type Error = String

Alias de tipo para mensajes de error.

newtype EvalSteps a

El tipo EvalSteps encapsula una computación que puede resultar en un error o en un valor, junto con una lista de mensajes de registro.

Constructors

EvalSteps

Fields

unEvalSteps :: (Either Error a, [String])

addTrace :: String -> EvalSteps ()

Función para agregar un mensaje de traza. Agrega un mensaje de registro a la computación EvalSteps.