Symplify

Documentation

automaticSymplify :: (MonadFail m, Alternative m) => PExpr -> m PExpr

simplifyProduct :: MonadFail m => [PExpr] -> m PExpr

simplifySum :: MonadFail m => [PExpr] -> m PExpr

simplifyDiv :: MonadFail m => PExpr -> PExpr -> m PExpr

simplifySub :: MonadFail m => PExpr -> PExpr -> m PExpr

simplifyPow :: MonadFail m => PExpr -> PExpr -> m PExpr

simplifyNegate :: MonadFail m => PExpr -> m PExpr

operands :: PExpr -> [PExpr]

mulByNeg :: PExpr -> Bool

freeOf :: PExpr -> PExpr -> Bool

linearForm :: MonadFail m => PExpr -> PExpr -> m (PExpr, PExpr)

numerator :: MonadFail m => PExpr -> m PExpr

denominator :: MonadFail m => PExpr -> m PExpr

simplifyFun :: (Alternative f, MonadFail f) => PExpr -> f PExpr

data TriBool

Tipo de dato que representa los valores de verdad, Verdadero, Falso y Desconocido.

Constructors

F
Falso

U
Desconocido

T
Verdadero

data PExpr

Constructors

Number Number

Mul [PExpr]

Add [PExpr]

Pow PExpr PExpr

Fun String [PExpr]

class BAlgebra a b c | a b -> c where

Clase que permite redefinir los operadores logicos para valores que no necesariamente son Booleanos, util para poder operar tanto con valores ternarios como booleanos.

Methods

(&&&) :: a -> b -> c infixr 3

Operador AND

(|||) :: a -> b -> c infixr 2

Operador OR

(!&&) :: a -> b -> c infixr 3

Operador NAND

(/||) :: a -> b -> c infixr 2

Operador XOR

pattern Symbol :: String -> PExpr

class Assumptions a where

Methods

isPositive :: a -> TriBool

isNegative :: a -> TriBool

isZero :: a -> TriBool

isEven :: a -> TriBool

isOdd :: a -> TriBool

isInteger :: a -> TriBool

pattern Pi :: PExpr

pattern Exp :: PExpr -> PExpr

pattern Log :: PExpr -> PExpr

pattern Sin :: PExpr -> PExpr

pattern Cos :: PExpr -> PExpr

pattern Tan :: PExpr -> PExpr

pattern Cot :: PExpr -> PExpr

pattern Sec :: PExpr -> PExpr

pattern Csc :: PExpr -> PExpr

pattern Asin :: PExpr -> PExpr

pattern Acos :: PExpr -> PExpr

pattern Atan :: PExpr -> PExpr

pattern Sinh :: PExpr -> PExpr

pattern Cosh :: PExpr -> PExpr

pattern Tanh :: PExpr -> PExpr

pattern Asinh :: PExpr -> PExpr

pattern Acosh :: PExpr -> PExpr

pattern Atanh :: PExpr -> PExpr

true :: TriBool -> Bool

Retorna True si el valor es Verdadero, False en caso contrario.

false :: TriBool -> Bool

Retorna True si el valor es Falso, False en caso contrario.

unknown :: TriBool -> Bool

Retorna True si el valor es Desconocido, False en caso contrario.

triBoolCase :: TriBool -> a -> a -> a -> a

"if then else" para valores de verdad ternarios

liftBool :: Bool -> TriBool

Convierte un valor de verdad booleano a uno ternario.

not3 :: BAlgebra b b c => b -> c

Operador NOT

xor3 :: (a -> TriBool) -> [a] -> TriBool

Aplica un predicado a una lista de valores y retorna el XOR ternario de todos

and3 :: (a -> TriBool) -> [a] -> TriBool

Aplica un predicado a una lista de valores y retorna el AND ternario de todos

or3 :: (a -> TriBool) -> [a] -> TriBool

Aplica un predicado a una lista de valores y retorna el OR ternario de todos