TriBool

Description

La logica clasica soporta dos valores de verdad, Verdadero y Falso. Sin embargo, hay situaciones en las que no se puede determinar si una proposicion es verdadera o falsa. Por ejemplo, la expresion 0^x evalua a aiferentes valores dependiendo del valor de x.

    0 ^ x = 0, si x > 0
    0 ^ x = Undefined: "Division por 0", si x <= 0

Pero el valor de x no siempre es conocido, por lo que el valor de verdad de x > 0 no siempre se puede inferir como verdadero o falso.

La logica ternaria aborda este problema, introduciendo un tercer valor de verdad, Desconocido (U). Este modulo proporciona una implementacion de la logica ternaria, que soporta los valores de verdad Verdadero (T), Falso (F) y Desconocido (U), ademas de una extensión de los operadores logicos para estos valores.

Mas información: https://en.wikipedia.org/wiki/Three-valued_logic

Ejemplo de uso:

import TriBool

-- and
T &&& T == T 
T &&& F == F
U &&& T == U

-- or
F ||| F = F 
F ||| U = U
T ||| U = T

Notar que las operaciones con valores desconocidos no siempre resultan en un valor desconocido.

Documentation

data TriBool

Tipo de dato que representa los valores de verdad, Verdadero, Falso y Desconocido.

Constructors

F
Falso

U
Desconocido

T
Verdadero

true :: TriBool -> Bool

Retorna True si el valor es Verdadero, False en caso contrario.

false :: TriBool -> Bool

Retorna True si el valor es Falso, False en caso contrario.

unknown :: TriBool -> Bool

Retorna True si el valor es Desconocido, False en caso contrario.

triBoolCase :: TriBool -> a -> a -> a -> a

"if then else" para valores de verdad ternarios

liftBool :: Bool -> TriBool

Convierte un valor de verdad booleano a uno ternario.

class BAlgebra a b c | a b -> c where

Clase que permite redefinir los operadores logicos para valores que no necesariamente son Booleanos, util para poder operar tanto con valores ternarios como booleanos.

Methods

(&&&) :: a -> b -> c infixr 3

Operador AND

(|||) :: a -> b -> c infixr 2

Operador OR

(!&&) :: a -> b -> c infixr 3

Operador NAND

(/||) :: a -> b -> c infixr 2

Operador XOR

not3 :: BAlgebra b b c => b -> c

Operador NOT

xor3 :: (a -> TriBool) -> [a] -> TriBool

Aplica un predicado a una lista de valores y retorna el XOR ternario de todos

and3 :: (a -> TriBool) -> [a] -> TriBool

Aplica un predicado a una lista de valores y retorna el AND ternario de todos

or3 :: (a -> TriBool) -> [a] -> TriBool

Aplica un predicado a una lista de valores y retorna el OR ternario de todos